added stock validation before saving a new venta

+++ src/main/java/xatal/sharedz/structures/PublicProductoVenta.java
@@ -3,14 +3,14 @@ package xatal.sharedz.structures;
 import xatal.sharedz.entities.ProductoVenta;
 
 public class PublicProductoVenta {
-    public int producto = -1;
+    public Long producto = -1L;
     public int cantidad = 0;
 
     public PublicProductoVenta() {
     }
 
     public PublicProductoVenta(ProductoVenta venta) {
-        this.producto = Math.toIntExact(venta.getProducto());
+        this.producto = venta.getProducto();
         this.cantidad = venta.getCantidad();
     }
 }
